extracted views from cart view
