+++ RiyadhSouq/Views/CartView.swift
@@ -37,90 +37,6 @@ struct CartView: View {
   }
 }
 
-struct CartItemRow: View {
-  let item: Product // Assuming Product is your data model
-  var body: some View {
-    HStack {
-      if let imageUrl = item.images.first, let url = URL(string: imageUrl) {
-        AsyncImage(url: url) { image in
-          image.resizable()
-        } placeholder: {
-          ProgressView()
-        }
-        .frame(width: 60, height: 60)
-        .cornerRadius(8)
-      }
-      VStack(alignment: .leading) {
-        Text(item.title)
-          .font(.headline)
-        Text("Price: $ \(item.price)")
-          .font(.subheadline)
-      }
-    }
-  }
-}
-
-struct PromoCodeInput: View {
-  @Binding var promoCode: String
-  @Binding var discountAmount: Int
-  @Binding var showInvalidPromoCodeAlert: Bool
-  var cartViewModel: CartViewModel
-  var body: some View {
-    HStack {
-      TextField("Enter promo code", text: $promoCode)
-        .textFieldStyle(RoundedBorderTextFieldStyle())
-        .padding()
-      Button("Apply") {
-        if promoCode.isEmpty {
-          discountAmount = 0
-        } else {
-          let discount = cartViewModel.applyPromoCode(promoCode)
-          if discount > 0 {
-            discountAmount = discount
-          } else {
-            discountAmount = 0
-            showInvalidPromoCodeAlert = true
-          }
-        }
-      }
-      .padding()
-    }
-    .alert(isPresented: $showInvalidPromoCodeAlert) {
-      Alert(title: Text("Invalid Promo Code"),
-            message: Text("The promo code you entered is not valid."),
-            dismissButton: .default(Text("OK")))
-    }
-  }
-}
-
-struct TotalPriceView: View {
-  let originalTotalPrice: Int
-  let discountAmount: Int
-  let discountedTotalPrice: Int
-  var body: some View {
-    HStack {
-      Spacer()
-      VStack(alignment: .trailing) {
-        if discountAmount > 0 {
-          Text("Original Price: $ \(originalTotalPrice)")
-            .strikethrough()
-            .font(.body)
-            .fontWeight(.bold)
-            .foregroundColor(.secondary)
-          Text("Discount: $ \(discountAmount)")
-            .font(.body)
-            .fontWeight(.bold)
-            .foregroundColor(.red)
-        }
-        Text("Total Price: $ \(discountedTotalPrice)")
-          .font(.headline)
-          .fontWeight(.bold)
-      }
-      .padding()
-    }
-  }
-}
-
 struct CartView_Previews: PreviewProvider {
   static var previews: some View {
     CartView(cartViewModel: CartViewModel())
